// src/lib/careerdna/types.ts
export interface RawPsychometricScores {
  extraversion: number;
  agreeableness: number;
  conscientiousness: number;
  neuroticism: number;
  openness: number;
  realistic: number;
  investigative: number;
  artistic: number;
  social: number;
  enterprising: number;
  conventional: number;
  numericalReasoning: number;
  verbalReasoning: number;
  logicalReasoning: number;
  stability: number;
  materialReward: number;
  learning: number;
  creativity: number;
  altruism: number;
  influence: number;
  recognition: number;
  autonomy: number;
  teamwork: number;
  intellectualChallenge: number;
  leadership: number;
  conflictResolution: number;
  teamDynamics: number;
  integrity: number;
  professionalism: number;
  accountability: number;
  clientManagement: number;
  decisionMaking: number;
  resilience: number;
  timeManagement: number;
  collaboration: number;
  learningAgility: number;
  proactivity: number;
  prioritization: number;
  stakeholderManagement: number;
  communication: number;
  selfAwareness: number;
  vark: number;
}